Updated Breadth First Traversal
I wasn't satisfied with the previous function and how I went about doing it. Especially since I know that I didn't do it the right way by checking every node on each level. So that's what I did. I created a separate function to that though and I managed to make it output the correct path.