Semi-completed A* Search
It does what it's supposed to do. It correctly prints the fringe as how I would have drawn it on paper. The issue is, it can't give me the proper path at the moment. I'd have to look on the graph itself to determine that. I have to draw it anyways.